perf(ingest): the preprocessor-dead walk was O(C^2) — llvm cold 202.14 -> 170.46 s CPU, map byte-identical
collectPreprocDeadRanges (src/preprocdead.h) read its children with ts_node_child( n, i ), which
restarts tree-sitter's child iterator from the first child on every call, so the walk was O(C^2) in
a node's child count — audit P1-0, the highest finding of the 2026-09-10 round. The repo already
states that rule and ships the cursor helper for it, but the helper lived inside ingest.cpp's unnamed
namespace (ingest_metrics.h) where preprocdead.h — compiled into slice.h too — could not reach it.
So the helper moves to src/infra/tschildren.h, gains an appendChildren form for the DFS-stack case,
and the walk uses it. Same child set, same left-to-right collection order, same reverse-child visit
order, same emitted ranges.

Why it hid: preprocdead.h short-circuits on src.find("#if") == npos, so Go/Python/JS corpora never
enter the walk and test/padscalecheck.sh's comment-flood fixture (which has no #if) cannot reach it.
Every C/C++ INCLUDE GUARD opens that gate and makes the guard's preproc_ifdef node one node whose
child list is the whole file.

A/B, one interleaved cold pair, llvm-project (2.9 GB, 182k files), same box, load ~12:

  arm   CPU (user+sys)   wall     max RSS    map
  pre       202.14 s    26.60 s   6.16 GB    43485 B
  new       170.46 s    18.99 s   6.23 GB    43485 B   byte-identical
  delta      -15.7%     -28.6%

Leaf attribution, 12 s sample of a cold llvm run (audit pre-figures vs this lane's post-figures):
  ts_node_child_iterator_next      62.99% -> 13.44% of busy
  collectPreprocDeadRanges (incl.) 56.67% ->  2.07% of busy
The 31.7 s realised is short of the 107 s the 56.67% share implies: that share was read from a 12 s
window of a 26 s run, and a leaf share is not a whole-run share. The A/B is the number to believe.

Generated fixture (include guard + N line comments + an #if 0/#else pair), pre binary:
  N        pre CPU   new CPU
  1000      0.01 s    0.01 s
  4000      0.09 s    0.01 s
  16000     1.24 s    0.01 s        124x
Isolating control — the identical 16000-line flood with the guard REMOVED (no #if text, so the walk
never runs) costs 0.02 s on the pre binary, i.e. the whole 1.24 s was this one loop.

Narrow-tree corpora are inside the box's noise band, in both directions: paired interleaved cold runs
gave ripwire's own tree +2.8% median (20 pairs) and the go corpus -2.1% median (12 pairs) — and go
contains no #if at all, so the walk provably never runs there. No claim is made either way.

Non-degradation — every one byte-identical, pre binary vs new, --no-cache:
  corpus                       map(--top-k=100000)  --for  --grep  --pack-task  --dead-code  --lint
  ripwire (this tree)                 ok              ok     ok        ok           ok         ok
  go (227 MB)                         ok              ok     ok        ok           ok         ok
  canyonraid48/canyon (C++)           ok              ok     ok        ok           ok         ok
  llvm-project (map)                  ok
Plus determinism (two cold runs cmp-equal) and xmllint --noout on the map.

Gate (written first, red before the code): test/preprocdeadscalecheck.sh. Arms (B) 62.0x CPU for 16x
the child width and (C) an include guard costing 62.0x the identical unguarded flood both FAILED
against the pre-change binary and PASS after; (A) asserts the dead-range set through --uses on the
flooded fixture (both halves: the #if 0 call absent, the live call present) plus determinism;
(D1)/(D2) hold byte-identity against RIPWIRE_REF_BIN on the generated fixtures and on five committed
C/C++ fixture trees; (E) shows every verdict and row reader able to fail. Its header carries the
FOLLOW-UP enumeration of the 52 surviving indexed ts_node_child( n, i ) sites in three trip classes —
10 unbounded (class 1, the same defect), 5 input-controlled but small (class 2), ~37 grammar-bounded
and correct as written (class 3). Not converted here: they live in files other lanes are editing.

Verified: preprocdeadscalecheck, padscalecheck, preproccondcheck, blindspotcheck, slicecheck,
includeprecisecheck, rustimportprecisecheck, loopconservationcheck, readmedriftcheck,
cacheidentitycheck, limitstablecheck, infraportcheck, includeanglecheck, selfcontainedcheck,
shellgateindexcheck, binoverridecheck, nodekindcheck, xmlwellformed, manifestcheck, gatecountcheck —
all ALL PASS. ASan (-fno-sanitize-recover=all, LSan with the committed suppressions) clean on the
generated fixtures, test/preproccondfix and ripwire's own tree, and the gate passes under it — the
cursor's ts_tree_cursor_delete lifetime is RAII-held by ChildCursor. --quality-delta gating=0
(5 minor new-symbol api-surface rows: the moved helper is now a named rw:: symbol instead of an
internal-linkage one — my footprint, deliberately not acked). Gate count regenerated to 587 by
docs/gatecount_build.py, never hand-written.
